Rearranged files and functions + added static
